cmake: fix libatomic checks
CMAKE_REQUIRED_FLAGS is not appropriate for libraries.
They are added before object files to the compiler command line arguments.
Depending on compiler settings it might decide that the library is not
actually needed because at that point it hasn't seen any object file
using a symbol from the library.
See "-Wl,--as-needed"