Allow usage of boost in different prefix
Right now, even if CMAKE_PREFIX_PATH for instance is set and points to
an up-to-date boost, CMake will just link to the system boost, as it is
never told where the right headers and libraries are.

A far more superior solution would be to require a CMake version 3.5+
and use the imported targets FindBoost provides (Boost::*).

See https://cmake.org/cmake/help/v3.5/module/FindBoost.html for more
information.